Comitando operação de porcentagem

+++ functions/digitandoNumeros.js
@@ -5,7 +5,8 @@ const botoes = document.querySelectorAll("[data-botao]");
 export const digitandoNumeros = (input) => {
   botoes.forEach((botao) => {
     botao.addEventListener("click", () => {
-      input.value += parseInt(botao.textContent);
+      input.value += parseFloat(botao.textContent);
+      
     });
   });
 };
